Revert to using older sysinfo when on 3.X
