+++ src/browser/modules/Stream/SysInfoFrame/LegacySysInfoFrame/sysinfo-utils.test.ts
